requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
loguru>=0.7.2
alibabacloud-dingtalk>=2.0.0
mcp-server>=0.1.0
sqlalchemy>=2.0.0
pydantic>=2.0.0

# OpenAI Agents
openai-agents>=0.0.14
